Add Rust + testing infrastructure for 20.04 upgrade
This commit adds the Rust toolchain and removes `blackbox` from
installation (given as there doesn't seem to be much use for it in my
regular work).

More importantly, this commit adds a way for me to test if my
configuration scripts work on Ubuntu versions besides 18.04 without
needing a physical machine. The tooling consists of a Docker setup with
systemd as the root process on a Ubuntu version of my choosing, allowing
me to run `sudo ./launch.sh` with different Salt versions and different
Ubuntu versions and debug issues there.

Finally, the default is now Salt v3003 going forward.